feat: getting codes into a pdf
added one function to get the content for the pdf and one to safe it. Also added a folder with the pdfs. As soon as you click on the generate code button it will generate the pdf. Only problem is you'll get an error message since it has to load the font for the pdf. Nevertheless it works